procfs kernel patch by Tony Espy Just wanted to let folks now that we've added a kernel patch to all of our current public kernels which fixes some problems with the current application life cycle due to PID mis-matches between Android and the Ubuntu container.
This patch allows code running on the Android side to properly lookup a
processes PID from the Ubuntu namespace, so that it both sides can refer
to the same process.

This is a short-term solution, however it unblocks a bug which prevents
applications from being launched from the command-line without manual
intervention to get the app to display.

If you're involved in porting Ubuntu Touch to a new device ( or have
already done so ), you'll need to apply this patch to your kernels,
which should result in better application behavior.

by Tony Espy